+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## [1.0.0-beta.2]
+### Added
+- Added support to manually set bulk size and enable/disable bulk dispatch
+- `loadConfigurations` and `loadConfigurationsFromJson` functions
+- Updated to use `requires_token`
+
+### Fixed
+- Issue with plain log requests fixed
+- Issue with error connection
+
 ## [1.0.0-beta.1]
 ### Changes
  - Initial release